Update for D
Sequential: remove a dependency for HashMap. Now it is based on built-in version
Parallel: slightly update the code that it will be inline with format of sequential version
Major part: changed the behavior of building option. Now the code use bounds-check and only "optimize" and "inline" options used in "release" mode. Previously bounds-checking were disabled by default of building in "release" mode.
Update names for output files of parallel version